feat: add saveAndUpdate method for less code repetition and refactor methods from model to dto

+++ src/main/java/org/example/schoolsystem/service/ClassService.java
@@ -11,6 +11,7 @@ import org.example.schoolsystem.model.TeacherModel;
 import org.example.schoolsystem.repository.ClassRepository;
 import org.example.schoolsystem.repository.StudentRepository;
 import org.example.schoolsystem.repository.TeacherRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -25,28 +26,59 @@ public class ClassService {
     private final StudentRepository studentRepository;
     private final TeacherRepository teacherRepository;
     private static final Logger LOGGER = org.apache.logging.log4j.LogManager.getLogger(ClassService.class);
+    private final ModelMapper modelMapper;
 
-    public ClassModel findClassById(UUID id){
+    public ClassDTO findClassById(UUID id) {
         LOGGER.info("Finding class by id: " + id);
-        return classRepository.findById(id).orElseThrow(() -> {
+        ClassModel classModel = modelMapper.map(classRepository.findById(id).orElseThrow(() -> {
             LOGGER.error("Class not found");
             return new NotFoundException("Class not found");
-        });
+        }), ClassModel.class);
+        return modelMapper.map(classModel, ClassDTO.class);
+    }
+
+    public ClassDTO findClassBySubject(String subject){
+        LOGGER.info("Finding class by subject: " + subject);
+        ClassModel classModel = modelMapper.map(classRepository.findBySubject(subject), ClassModel.class);
+
+        if (classModel == null) {
+            LOGGER.error("Class not found");
+            throw new NotFoundException("Class not found");
+        }
+
+        LOGGER.info("Class found successfully");
+        return modelMapper.map(classModel, ClassDTO.class);
     }
 
 
-    public List<ClassModel> findAllClasses(){
+    public List<ClassDTO> findAllClasses(){
         LOGGER.info("Finding all classes");
-        return classRepository.findAll();
+        ClassModel classModel = modelMapper.map(classRepository.findAll(), ClassModel.class);
+        if (classModel == null) {
+            LOGGER.error("Class not found");
+            throw new NotFoundException("Class not found");
+        }
+        return modelMapper.map(classRepository.findAll(), List.class);
     }
 
     @Transactional
-    public ClassModel saveClass(ClassDTO classDTO) {
-        ClassModel classModel = new ClassModel();
+    public ClassDTO saveAndUpdateClass(ClassDTO classDTO) {
+        ClassModel classModel;
+
+        if (classDTO.getId() != null && classRepository.existsById(classDTO.getId())) {
+            classModel = classRepository.findById(classDTO.getId())
+                    .orElseThrow(() -> new NotFoundException("Class not found"));
+        } else {
+            classModel = new ClassModel();
+        }
+
+        // Set teacher
         TeacherModel teacher = teacherRepository.findById(classDTO.getTeacherId())
                 .orElseThrow(() -> new NotFoundException("Teacher not found"));
-
         classModel.setTeacher(teacher);
+
+        // Set subject
+        classModel.setSubject(classDTO.getSubject());
         // Set students
         Set<StudentModel> students = classDTO.getStudents().stream()
                 .map(UUID::fromString)
@@ -57,17 +89,31 @@ public class ClassService {
         classModel.setStudents(students);
         classRepository.save(classModel);
         LOGGER.info("Class saved successfully");
-        return classModel;
+        return modelMapper.map(classModel, ClassDTO.class);
     }
 
-    public ClassModel deleteClassById(UUID id) {
-        ClassModel classModel = classRepository.findById(id).orElseThrow(() -> {
+    @Transactional
+    public ClassDTO saveClass(ClassDTO classDTO){
+        LOGGER.info("Saving class...");
+        return saveAndUpdateClass(classDTO);
+    }
+
+    @Transactional
+    public ClassDTO updateClass(ClassDTO classDTO){
+        LOGGER.info("Updating class...");
+        return saveAndUpdateClass(classDTO);
+    }
+
+    public void deleteClassById(UUID id) {
+        LOGGER.info("Deleting class...");
+
+        if (!classRepository.existsById(id)){
             LOGGER.error("Class not found");
-            return new NotFoundException("Class not found for deletion");
-        });
+            throw new NotFoundException("Class not found");
+        }
+
         classRepository.deleteById(id);
         LOGGER.info("Class deleted successfully");
-        return classModel;
     }
 
 }
